Add gamma correctio example
The example is not the best, better version should be created. The main
issue of the current version is that it does not clearly show what is
wrong/right (the perceptually smooth gradient is good? not it's not)
unless the person seeing it know how gamma correction works and why is
it good. We should rather change the example to something similarly
simple as the gradient but with clear right/wrong scenario where one
image is clearly visually superior to the other. For now this may
suffice.